Whitepapers template and functionality finished ✔️
