+++ .vscode/settings.json
@@ -101,6 +101,7 @@
     "jank",
     "keegan",
     "keychain",
+    "Kitko",
     "lims",
     "linkedin",
     "makemigrations",
