+++ tools/data/manage_analyses.py
@@ -1,5 +1,5 @@
 """
-Manage Regulations | Cannlytics Website
+Manage Analyses | Cannlytics Website
 Copyright (c) 2021-2022 Cannlytics
 
 Authors: Keegan Skeate <keegan@cannlytics.com>
@@ -12,13 +12,13 @@ Command-line examples:
     Get and save data.
 
     ```
-    python tools/manage_regulations.py get_regulation_data
+    python tools/data/manage_analyses.py get_analysis_data
     ```
 
     Upload data.
 
     ```
-    python tools/manage_regulations.py upload_regulation_data
+    python tools/data/manage_analyses.py upload_analysis_data
     ```
 """
 # Standard imports.
@@ -32,33 +32,33 @@ from dotenv import dotenv_values
 from data_management import get_data, upload_data
 
 # Define references.
-FILENAME = 'regulations.json'
+FILENAME = 'analyses.json'
 DOC = 'public/data'
-REF = f'{DOC}/regulations'
-ID = 'state'
+REF = f'{DOC}/analyses'
+ID = 'analysis_id'
 
 
-def get_regulation_data():
-    """Get regulation data from Firestore."""
+def get_analysis_data():
+    """Get analysis data from Firestore."""
     try:
         return get_data(REF, datafile=f'.datasets/{FILENAME}')
     except FileNotFoundError:
-        return get_data(REF, datafile=f'../.datasets/{FILENAME}')
+        return get_data(REF, datafile=f'../../.datasets/{FILENAME}')
 
 
-def upload_regulation_data():
-    """Upload regulation data from local `.datasets`."""
+def upload_analysis_data():
+    """Upload analysis data from local `.datasets`."""
     try:
         upload_data(f'.datasets/{FILENAME}', REF, id_key=ID, stats_doc=DOC)
     except FileNotFoundError:
-        upload_data(f'../.datasets/{FILENAME}', REF, id_key=ID, stats_doc=DOC)
+        upload_data(f'../../.datasets/{FILENAME}', REF, id_key=ID, stats_doc=DOC)
 
 
 if __name__ == '__main__':
 
     # Set credentials.
     try:
-        config = dotenv_values('../.env')
+        config = dotenv_values('../../.env')
         credentials = config['GOOGLE_APPLICATION_CREDENTIALS']
     except KeyError:
         config = dotenv_values('.env')
