+++ tools/data/manage_regulations.py
@@ -1,10 +1,10 @@
 """
-Manage Sponsorships | Cannlytics Website
+Manage Regulations | Cannlytics Website
 Copyright (c) 2021-2022 Cannlytics
 
 Authors: Keegan Skeate <keegan@cannlytics.com>
 Created: 12/26/2021
-Updated: 1/3/2022
+Updated: 1/9/2022
 License: MIT License <https://github.com/cannlytics/cannlytics-website/blob/main/LICENSE>
 
 Command-line examples:
@@ -12,13 +12,13 @@ Command-line examples:
     Get and save data.
 
     ```
-    python tools/manage_sponsorships.py get_sponsorship_data
+    python tools/data/manage_regulations.py get_regulation_data
     ```
 
     Upload data.
 
     ```
-    python tools/manage_sponsorships.py upload_sponsorship_data
+    python tools/data/manage_regulations.py upload_regulation_data
     ```
 """
 # Standard imports.
@@ -32,32 +32,33 @@ from dotenv import dotenv_values
 from data_management import get_data, upload_data
 
 # Define references.
-FILENAME = 'sponsorships.json'
-DOC = 'public/subscriptions'
-REF = f'{DOC}/sponsorships'
+FILENAME = 'regulations.json'
+DOC = 'public/data'
+REF = f'{DOC}/regulations'
+ID = 'state'
 
 
-def get_sponsorship_data():
-    """Get sponsorship data from Firestore."""
+def get_regulation_data():
+    """Get regulation data from Firestore."""
     try:
         return get_data(REF, datafile=f'.datasets/{FILENAME}')
     except FileNotFoundError:
-        return get_data(REF, datafile=f'../.datasets/{FILENAME}')
+        return get_data(REF, datafile=f'../../.datasets/{FILENAME}')
 
 
-def upload_sponsorship_data():
-    """Upload sponsorship data from local `.datasets`."""
+def upload_regulation_data():
+    """Upload regulation data from local `.datasets`."""
     try:
-        upload_data(f'.datasets/{FILENAME}', REF, stats_doc=DOC)
+        upload_data(f'.datasets/{FILENAME}', REF, id_key=ID, stats_doc=DOC)
     except FileNotFoundError:
-        upload_data(f'../.datasets/{FILENAME}', REF, stats_doc=DOC)
+        upload_data(f'../../.datasets/{FILENAME}', REF, id_key=ID, stats_doc=DOC)
 
 
 if __name__ == '__main__':
 
     # Set credentials.
     try:
-        config = dotenv_values('../.env')
+        config = dotenv_values('../../.env')
         credentials = config['GOOGLE_APPLICATION_CREDENTIALS']
     except KeyError:
         config = dotenv_values('.env')
