+++ tools/data/manage_sponsorships.py
@@ -1,10 +1,10 @@
 """
-Manage Analyses | Cannlytics Website
+Manage Sponsorships | Cannlytics Website
 Copyright (c) 2021-2022 Cannlytics
 
 Authors: Keegan Skeate <keegan@cannlytics.com>
 Created: 12/26/2021
-Updated: 1/9/2022
+Updated: 1/3/2022
 License: MIT License <https://github.com/cannlytics/cannlytics-website/blob/main/LICENSE>
 
 Command-line examples:
@@ -12,13 +12,13 @@ Command-line examples:
     Get and save data.
 
     ```
-    python tools/manage_analyses.py get_analysis_data
+    python tools/data/manage_sponsorships.py get_sponsorship_data
     ```
 
     Upload data.
 
     ```
-    python tools/manage_analyses.py upload_analysis_data
+    python tools/data/manage_sponsorships.py upload_sponsorship_data
     ```
 """
 # Standard imports.
@@ -32,33 +32,32 @@ from dotenv import dotenv_values
 from data_management import get_data, upload_data
 
 # Define references.
-FILENAME = 'analyses.json'
-DOC = 'public/data'
-REF = f'{DOC}/analyses'
-ID = 'analysis_id'
+FILENAME = 'sponsorships.json'
+DOC = 'public/subscriptions'
+REF = f'{DOC}/sponsorships'
 
 
-def get_analysis_data():
-    """Get analysis data from Firestore."""
+def get_sponsorship_data():
+    """Get sponsorship data from Firestore."""
     try:
         return get_data(REF, datafile=f'.datasets/{FILENAME}')
     except FileNotFoundError:
-        return get_data(REF, datafile=f'../.datasets/{FILENAME}')
+        return get_data(REF, datafile=f'../../.datasets/{FILENAME}')
 
 
-def upload_analysis_data():
-    """Upload analysis data from local `.datasets`."""
+def upload_sponsorship_data():
+    """Upload sponsorship data from local `.datasets`."""
     try:
-        upload_data(f'.datasets/{FILENAME}', REF, id_key=ID, stats_doc=DOC)
+        upload_data(f'.datasets/{FILENAME}', REF, stats_doc=DOC)
     except FileNotFoundError:
-        upload_data(f'../.datasets/{FILENAME}', REF, id_key=ID, stats_doc=DOC)
+        upload_data(f'../../.datasets/{FILENAME}', REF, stats_doc=DOC)
 
 
 if __name__ == '__main__':
 
     # Set credentials.
     try:
-        config = dotenv_values('../.env')
+        config = dotenv_values('../../.env')
         credentials = config['GOOGLE_APPLICATION_CREDENTIALS']
     except KeyError:
         config = dotenv_values('.env')
