+++ tools/data/manage_verifications.py
@@ -1,9 +1,9 @@
 """
-Manage Subscriptions | Cannlytics Website
+Manage Verifications | Cannlytics Website
 Copyright (c) 2021-2022 Cannlytics
 
 Authors: Keegan Skeate <keegan@cannlytics.com>
-Created: 12/26/2021
+Created: 11/15/2021
 Updated: 12/26/2021
 License: MIT License <https://github.com/cannlytics/cannlytics-website/blob/main/LICENSE>
 
@@ -12,13 +12,13 @@ Command-line examples:
     Get and save data.
 
     ```
-    python tools/manage_subscriptions.py get_subscription_data
+    python tools/data/manage_verifications.py get_verification_data
     ```
 
     Upload data.
 
     ```
-    python tools/manage_subscriptions.py upload_subscription_data
+    python tools/data/manage_verifications.py upload_verification_data
     ```
 """
 # Standard imports
@@ -33,7 +33,7 @@ from data_management import get_data, upload_data
 
 # Set credentials.
 try:
-    config = dotenv_values('../.env')
+    config = dotenv_values('../../.env')
     credentials = config['GOOGLE_APPLICATION_CREDENTIALS']
 except KeyError:
     config = dotenv_values('.env')
@@ -41,23 +41,23 @@ except KeyError:
 os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = credentials
 
 
-def get_subscription_data():
-    """Get subscription data from Firestore."""
-    ref = 'public/subscriptions/subscription_plans'
+def get_verification_data():
+    """Get verification data from Firestore."""
+    ref = 'public/verifications/verification_data'
     try:
-        return get_data(REF, datafile='.datasets/subscriptions.json')
+        return get_data(REF, datafile='.datasets/verifications.json')
     except FileNotFoundError:
-        return get_data(REF, datafile='../.datasets/subscriptions.json')
+        return get_data(REF, datafile='../../.datasets/verifications.json')
 
 
-def upload_subscription_data():
-    """Upload subscription data from local `.datasets`."""
-    ref = 'public/subscriptions/subscription_plans'
-    stats_doc = 'public/subscriptions'
+def upload_verification_data():
+    """Upload verification data from local `.datasets`."""
+    ref = 'public/verifications/verification_data'
+    stats_doc = 'public/verifications'
     try:
-        upload_data('.datasets/subscriptions.json', ref, stats_doc=stats_doc)
+        upload_data('.datasets/verifications.json', ref, stats_doc=stats_doc)
     except FileNotFoundError:
-        upload_data('../.datasets/subscriptions.json', ref, stats_doc=stats_doc)
+        upload_data('../../.datasets/verifications.json', ref, stats_doc=stats_doc)
 
 
 if __name__ == '__main__':
