+++ website/templates/website/pages/articles/article.html
@@ -8,19 +8,6 @@
   Created: 1/14/2022
   Updated: 1/14/2022
   License: MIT License <https://github.com/cannlytics/cannlytics-website/blob/main/LICENSE>
-
-  FIXME: Determine what parameters need to be passed and how.
-  Fields:
-    video_id (str): The unique ID of the video.
-    video_data (dict): A dictionary of video data, including the following fields:
-      `cover_image_url`, `description`,  `premium`,  `published`, `title`, `youtube_id`.
-    more_videos (list): A list of video data (dict) for 3 other random videos.
-    recent_videos (lsit): A list of video data (dict) for the 3 most recent videos.
-    user_id (str): The unique ID of any signed-in user.
-    premium_subscription (bool): Whether or not the user is subscribed to premium content.
-
-  TODO:
-    1. Add resources, such as links to the Github repository and downloadable files.
 -->
 {% load title_case %}
 <section class="container">
