+++ website/templates/website/pages/articles/articles.html
@@ -42,14 +42,14 @@
           </h3>
         </a>
         <h6 class="blog-post-meta text-secondary">
-          <small class="serif">
+          <small class="serif text-dark">
             {{ item.published_at_formatted }} by
             <a class="serif text-dark" href="{{ item.author_url }}">
               {{ item.author }}
             </a>
           </small>
         </h6>
-        <p class="fst-italic">
+        <p class="fst-italic text-dark">
           {{ item.excerpt|safe }}
         </p>
         <a
